// frontend/src/generated/orval.schemas.ts
export type AdvancedUsersProfileUpdateUsername = string | null;

export type AdvancedUsersProfileUpdateAvatarUrl = string | null;

export type AdvancedUsersProfileUpdateAddress = string | null;

export type AdvancedUsersProfileUpdateCountry = string | null;

export type AdvancedUsersProfileUpdatePhoneNumber = string | null;

export interface AdvancedUsersProfileUpdate {
  user_id: number;
  username?: AdvancedUsersProfileUpdateUsername;
  avatar_url?: AdvancedUsersProfileUpdateAvatarUrl;
  address?: AdvancedUsersProfileUpdateAddress;
  country?: AdvancedUsersProfileUpdateCountry;
  phone_number?: AdvancedUsersProfileUpdatePhoneNumber;
}